Create new files for each method
