+++ bin/methods/added_msg.rb
@@ -1,0 +1,5 @@
+def added_msg(input)
+    clear
+    puts "#{input} created successfully ✅"
+    puts
+end
